# perf(sql): support advanced aggregate functions in parallel GROUP BY