--- a/core/src/main/java/io/questdb/PropServerConfiguration.java
+++ b/core/src/main/java/io/questdb/PropServerConfiguration.java
@@ -266,6 +266,7 @@ public class PropServerConfiguration implements ServerConfiguration {
     private final int sqlExpressionPoolCapacity;
     private final double sqlFastMapLoadFactor;
     private final int sqlFloatToStrCastScale;
+    private final int sqlGroupByAllocatorChunkSize;
     private final int sqlGroupByMapCapacity;
     private final int sqlGroupByPoolCapacity;
     private final int sqlHashJoinLightValueMaxPages;
@@ -923,6 +924,7 @@ public class PropServerConfiguration implements ServerConfiguration {
             this.sqlDoubleToStrCastScale = getInt(properties, env, PropertyKey.CAIRO_SQL_DOUBLE_CAST_SCALE, 12);
             this.sqlFloatToStrCastScale = getInt(properties, env, PropertyKey.CAIRO_SQL_FLOAT_CAST_SCALE, 4);
             this.sqlGroupByMapCapacity = getInt(properties, env, PropertyKey.CAIRO_SQL_GROUPBY_MAP_CAPACITY, 1024);
+            this.sqlGroupByAllocatorChunkSize = getIntSize(properties, env, PropertyKey.CAIRO_SQL_GROUPBY_ALLOCATOR_CHUNK_SIZE, 32 * 1024);
             this.sqlGroupByPoolCapacity = getInt(properties, env, PropertyKey.CAIRO_SQL_GROUPBY_POOL_CAPACITY, 1024);
             this.sqlMaxSymbolNotEqualsCount = getInt(properties, env, PropertyKey.CAIRO_SQL_MAX_SYMBOL_NOT_EQUALS_COUNT, 100);
             this.sqlBindVariablePoolSize = getInt(properties, env, PropertyKey.CAIRO_SQL_BIND_VARIABLE_POOL_SIZE, 8);
@@ -2045,6 +2047,11 @@ public class PropServerConfiguration implements ServerConfiguration {
         @Override
         public int getFloatToStrCastScale() {
             return sqlFloatToStrCastScale;
+        }
+
+        @Override
+        public int getGroupByAllocatorChunkSize() {
+            return sqlGroupByAllocatorChunkSize;
         }
 
         @Override