--- a/core/src/main/java/io/questdb/PropertyKey.java
+++ b/core/src/main/java/io/questdb/PropertyKey.java
@@ -118,6 +118,7 @@ public enum PropertyKey implements ConfigPropertyKey {
     CAIRO_SQL_FLOAT_CAST_SCALE("cairo.sql.float.cast.scale"),
     CAIRO_SQL_GROUPBY_MAP_CAPACITY("cairo.sql.groupby.map.capacity"),
     CAIRO_SQL_GROUPBY_POOL_CAPACITY("cairo.sql.groupby.pool.capacity"),
+    CAIRO_SQL_GROUPBY_ALLOCATOR_CHUNK_SIZE("cairo.sql.groupby.allocator.chunk.size"),
     CAIRO_SQL_MAX_SYMBOL_NOT_EQUALS_COUNT("cairo.sql.max.symbol.not.equals.count"),
     CAIRO_SQL_BIND_VARIABLE_POOL_SIZE("cairo.sql.bind.variable.pool.size"),
     CAIRO_SQL_QUERY_REGISTRY_POOL_SIZE("cairo.sql.query.registry.pool.size"),