--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -171,6 +171,8 @@ public interface CairoConfiguration {
 
     int getFloatToStrCastScale();
 
+    int getGroupByAllocatorChunkSize();
+
     int getGroupByMapCapacity();
 
     int getGroupByMergeShardQueueCapacity();