--- a/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfiguration.java
@@ -171,7 +171,9 @@ public interface CairoConfiguration {
 
     int getFloatToStrCastScale();
 
-    int getGroupByAllocatorChunkSize();
+    long getGroupByAllocatorDefaultChunkSize();
+
+    long getGroupByAllocatorMaxChunkSize();
 
     int getGroupByMapCapacity();