--- a/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
+++ b/core/src/main/java/io/questdb/cairo/CairoConfigurationWrapper.java
@@ -271,6 +271,11 @@ public class CairoConfigurationWrapper implements CairoConfiguration {
     }
 
     @Override
+    public int getGroupByAllocatorChunkSize() {
+        return delegate.getGroupByAllocatorChunkSize();
+    }
+
+    @Override
     public int getGroupByMapCapacity() {
         return delegate.getGroupByMapCapacity();
     }