--- a/core/src/main/java/io/questdb/cairo/map/FastMap.java
+++ b/core/src/main/java/io/questdb/cairo/map/FastMap.java
@@ -74,6 +74,7 @@ import org.jetbrains.annotations.TestOnly;
  */
 public class FastMap implements Map, Reopenable {
 
+    static final long VAR_KEY_HEADER_SIZE = 4;
     private static final long MAX_HEAP_SIZE = (Integer.toUnsignedLong(-1) - 1) << 3;
     private static final int MIN_INITIAL_CAPACITY = 128;
     private final FastMapCursor cursor;
@@ -81,9 +82,9 @@ public class FastMap implements Map, Reopenable {
     private final int initialKeyCapacity;
     private final int initialPageSize;
     private final BaseKey key;
-    private final int keyOffset;
+    private final long keyOffset;
     // Set to -1 when key is var-size.
-    private final int keySize;
+    private final long keySize;
     private final int listMemoryTag;
     private final double loadFactor;
     private final int maxResizes;
@@ -93,7 +94,7 @@ public class FastMap implements Map, Reopenable {
     private final FastMapValue value2;
     private final FastMapValue value3;
     private final int valueColumnCount;
-    private final int valueSize;
+    private final long valueSize;
     private long capacity;
     private int free;
     private long heapLimit; // Heap memory limit pointer.
@@ -171,7 +172,7 @@ public class FastMap implements Map, Reopenable {
         this.maxResizes = maxResizes;
 
         final int keyColumnCount = keyTypes.getColumnCount();
-        int keySize = 0;
+        long keySize = 0;
         for (int i = 0; i < keyColumnCount; i++) {
             final int columnType = keyTypes.getColumnType(i);
             final int size = ColumnType.sizeOf(columnType);
@@ -187,9 +188,9 @@ public class FastMap implements Map, Reopenable {
         // Reserve 4 bytes for key length in case of var-size keys.
         keyOffset = keySize != -1 ? 0 : Integer.BYTES;
 
-        int valueOffset = 0;
+        long valueOffset = 0;
         long[] valueOffsets = null;
-        int valueSize = 0;
+        long valueSize = 0;
         if (valueTypes != null) {
             valueColumnCount = valueTypes.getColumnCount();
             valueOffsets = new long[valueColumnCount];
@@ -325,7 +326,7 @@ public class FastMap implements Map, Reopenable {
 
     @Override
     public MapValue valueAt(long startAddress) {
-        int keySize = this.keySize;
+        long keySize = this.keySize;
         if (keySize == -1) {
             keySize = Unsafe.getUnsafe().getInt(startAddress);
         }
@@ -369,7 +370,7 @@ public class FastMap implements Map, Reopenable {
         assert keySize >= 0;
         setKeyCapacity(size + srcMap.size);
 
-        int len = keySize + valueSize;
+        long len = keySize + valueSize;
         long alignedLen = Bytes.align8b(len);
 
         OUTER:
@@ -448,7 +449,7 @@ public class FastMap implements Map, Reopenable {
             }
 
             assert free > 0;
-            int len = keyOffset + srcKeySize + valueSize;
+            long len = keyOffset + srcKeySize + valueSize;
             if (kPos + len > heapLimit) {
                 resize(len);
             }
@@ -461,7 +462,7 @@ public class FastMap implements Map, Reopenable {
         }
     }
 
-    private FastMapValue probe0(BaseKey keyWriter, int index, int hashCode, int keySize, FastMapValue value) {
+    private FastMapValue probe0(BaseKey keyWriter, int index, int hashCode, long keySize, FastMapValue value) {
         long offset;
         while ((offset = getOffset(offsets, index = (++index & mask))) > -1) {
             if (hashCode == getHashCode(offsets, index) && keyWriter.eq(offset)) {
@@ -472,7 +473,7 @@ public class FastMap implements Map, Reopenable {
         return asNew(keyWriter, index, hashCode, value);
     }
 
-    private FastMapValue probeReadOnly(BaseKey keyWriter, int index, int hashCode, int keySize, FastMapValue value) {
+    private FastMapValue probeReadOnly(BaseKey keyWriter, int index, int hashCode, long keySize, FastMapValue value) {
         long offset;
         while ((offset = getOffset(offsets, index = (++index & mask))) > -1) {
             if (hashCode == getHashCode(offsets, index) && keyWriter.eq(offset)) {
@@ -516,7 +517,7 @@ public class FastMap implements Map, Reopenable {
         keyCapacity = newKeyCapacity;
     }
 
-    private void resize(int size) {
+    private void resize(long size) {
         if (nResizes < maxResizes) {
             nResizes++;
             long kCapacity = (heapLimit - heapStart) << 1;
@@ -550,11 +551,11 @@ public class FastMap implements Map, Reopenable {
         return value.of(startAddress, valueAddress, heapLimit, newValue);
     }
 
-    int keySize() {
+    long keySize() {
         return keySize;
     }
 
-    int valueSize() {
+    long valueSize() {
         return valueSize;
     }
 
@@ -569,7 +570,7 @@ public class FastMap implements Map, Reopenable {
 
         @Override
         public MapValue createValue() {
-            int keySize = commit();
+            long keySize = commit();
             // calculate hash remembering "key" structure
             // [ key size | key block | value block ]
             int hashCode = hash();
@@ -578,7 +579,7 @@ public class FastMap implements Map, Reopenable {
 
         @Override
         public MapValue createValue(int hashCode) {
-            int keySize = commit();
+            long keySize = commit();
             return createValue(keySize, hashCode);
         }
 
@@ -613,7 +614,7 @@ public class FastMap implements Map, Reopenable {
             // no-op
         }
 
-        private MapValue createValue(int keySize, int hashCode) {
+        private MapValue createValue(long keySize, int hashCode) {
             int index = hashCode & mask;
             long offset = getOffset(offsets, index);
             if (offset < 0) {
@@ -626,7 +627,7 @@ public class FastMap implements Map, Reopenable {
         }
 
         private MapValue findValue(FastMapValue value) {
-            int keySize = commit();
+            long keySize = commit();
             int hashCode = hash();
             int index = hashCode & mask;
             long offset = getOffset(offsets, index);
@@ -641,13 +642,13 @@ public class FastMap implements Map, Reopenable {
             }
         }
 
-        protected void checkSize(int requiredKeySize) {
+        protected void checkSize(long requiredKeySize) {
             if (appendAddress + requiredKeySize + valueSize > heapLimit) {
                 resize(requiredKeySize);
             }
         }
 
-        abstract void copyFromRawKey(long ptr, int size);
+        abstract void copyFromRawKey(long ptr, long size);
 
         protected abstract boolean eq(long offset);
     }
@@ -655,7 +656,7 @@ public class FastMap implements Map, Reopenable {
     class FixedSizeKey extends BaseKey {
 
         @Override
-        public int commit() {
+        public long commit() {
             return keySize;
         }
 
@@ -683,23 +684,23 @@ public class FastMap implements Map, Reopenable {
 
         @Override
         public void putBool(boolean value) {
-            assert appendAddress + Byte.BYTES <= heapLimit;
+            assert appendAddress + 1L <= heapLimit;
             Unsafe.getUnsafe().putByte(appendAddress, (byte) (value ? 1 : 0));
-            appendAddress += Byte.BYTES;
+            appendAddress += 1L;
         }
 
         @Override
         public void putByte(byte value) {
-            assert appendAddress + Byte.BYTES <= heapLimit;
+            assert appendAddress + 1L <= heapLimit;
             Unsafe.getUnsafe().putByte(appendAddress, value);
-            appendAddress += Byte.BYTES;
+            appendAddress += 1L;
         }
 
         @Override
         public void putChar(char value) {
-            assert appendAddress + Character.BYTES <= heapLimit;
+            assert appendAddress + 2L <= heapLimit;
             Unsafe.getUnsafe().putChar(appendAddress, value);
-            appendAddress += Character.BYTES;
+            appendAddress += 2L;
         }
 
         @Override
@@ -709,55 +710,55 @@ public class FastMap implements Map, Reopenable {
 
         @Override
         public void putDouble(double value) {
-            assert appendAddress + Double.BYTES <= heapLimit;
+            assert appendAddress + 8L <= heapLimit;
             Unsafe.getUnsafe().putDouble(appendAddress, value);
-            appendAddress += Double.BYTES;
+            appendAddress += 8L;
         }
 
         @Override
         public void putFloat(float value) {
-            assert appendAddress + Float.BYTES <= heapLimit;
+            assert appendAddress + 4L <= heapLimit;
             Unsafe.getUnsafe().putFloat(appendAddress, value);
-            appendAddress += Float.BYTES;
+            appendAddress += 4L;
         }
 
         @Override
         public void putInt(int value) {
-            assert appendAddress + Integer.BYTES <= heapLimit;
+            assert appendAddress + 4L <= heapLimit;
             Unsafe.getUnsafe().putInt(appendAddress, value);
-            appendAddress += Integer.BYTES;
+            appendAddress += 4L;
         }
 
         @Override
         public void putLong(long value) {
-            assert appendAddress + Long.BYTES <= heapLimit;
+            assert appendAddress + 8L <= heapLimit;
             Unsafe.getUnsafe().putLong(appendAddress, value);
-            appendAddress += Long.BYTES;
+            appendAddress += 8L;
         }
 
         @Override
         public void putLong128(long lo, long hi) {
-            assert appendAddress + 16 <= heapLimit;
+            assert appendAddress + 16L <= heapLimit;
             Unsafe.getUnsafe().putLong(appendAddress, lo);
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES, hi);
-            appendAddress += 16;
+            appendAddress += 16L;
         }
 
         @Override
         public void putLong256(Long256 value) {
-            assert appendAddress + Long256.BYTES <= heapLimit;
+            assert appendAddress + 32L <= heapLimit;
             Unsafe.getUnsafe().putLong(appendAddress, value.getLong0());
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES, value.getLong1());
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES * 2, value.getLong2());
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES * 3, value.getLong3());
-            appendAddress += Long256.BYTES;
+            appendAddress += 32L;
         }
 
         @Override
         public void putShort(short value) {
-            assert appendAddress + Short.BYTES <= heapLimit;
+            assert appendAddress + 2L <= heapLimit;
             Unsafe.getUnsafe().putShort(appendAddress, value);
-            appendAddress += Short.BYTES;
+            appendAddress += 2L;
         }
 
         @Override
@@ -781,7 +782,7 @@ public class FastMap implements Map, Reopenable {
         }
 
         @Override
-        void copyFromRawKey(long srcPtr, int srcSize) {
+        void copyFromRawKey(long srcPtr, long srcSize) {
             assert srcSize == keySize;
             Vect.memcpy(appendAddress, srcPtr, srcSize);
             appendAddress += srcSize;
@@ -794,11 +795,12 @@ public class FastMap implements Map, Reopenable {
     }
 
     class VarSizeKey extends BaseKey {
-        private int len;
+        private long len;
 
         @Override
-        public int commit() {
-            Unsafe.getUnsafe().putInt(startAddress, len = (int) (appendAddress - startAddress - keyOffset));
+        public long commit() {
+            len = appendAddress - startAddress - keyOffset;
+            Unsafe.getUnsafe().putInt(startAddress, (int) len);
             return len;
         }
 
@@ -818,7 +820,7 @@ public class FastMap implements Map, Reopenable {
             if (value == null) {
                 putVarSizeNull();
             } else {
-                long len = value.length() + Integer.BYTES;
+                long len = value.length() + 4L;
                 if (len > Integer.MAX_VALUE) {
                     throw CairoException.nonCritical().put("binary column is too large");
                 }
@@ -833,23 +835,23 @@ public class FastMap implements Map, Reopenable {
 
         @Override
         public void putBool(boolean value) {
-            checkSize(1);
+            checkSize(1L);
             Unsafe.getUnsafe().putByte(appendAddress, (byte) (value ? 1 : 0));
             appendAddress += 1;
         }
 
         @Override
         public void putByte(byte value) {
-            checkSize(1);
+            checkSize(1L);
             Unsafe.getUnsafe().putByte(appendAddress, value);
-            appendAddress += 1;
+            appendAddress += 1L;
         }
 
         @Override
         public void putChar(char value) {
-            checkSize(Character.BYTES);
+            checkSize(2L);
             Unsafe.getUnsafe().putChar(appendAddress, value);
-            appendAddress += Character.BYTES;
+            appendAddress += 2L;
         }
 
         @Override
@@ -859,55 +861,55 @@ public class FastMap implements Map, Reopenable {
 
         @Override
         public void putDouble(double value) {
-            checkSize(Double.BYTES);
+            checkSize(8L);
             Unsafe.getUnsafe().putDouble(appendAddress, value);
-            appendAddress += Double.BYTES;
+            appendAddress += 8L;
         }
 
         @Override
         public void putFloat(float value) {
-            checkSize(Float.BYTES);
+            checkSize(4L);
             Unsafe.getUnsafe().putFloat(appendAddress, value);
-            appendAddress += Float.BYTES;
+            appendAddress += 4L;
         }
 
         @Override
         public void putInt(int value) {
-            checkSize(Integer.BYTES);
+            checkSize(4L);
             Unsafe.getUnsafe().putInt(appendAddress, value);
-            appendAddress += Integer.BYTES;
+            appendAddress += 4L;
         }
 
         @Override
         public void putLong(long value) {
-            checkSize(Long.BYTES);
+            checkSize(8L);
             Unsafe.getUnsafe().putLong(appendAddress, value);
-            appendAddress += Long.BYTES;
+            appendAddress += 8L;
         }
 
         @Override
         public void putLong128(long lo, long hi) {
-            checkSize(16);
+            checkSize(16L);
             Unsafe.getUnsafe().putLong(appendAddress, lo);
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES, hi);
-            appendAddress += 16;
+            appendAddress += 16L;
         }
 
         @Override
         public void putLong256(Long256 value) {
-            checkSize(Long256.BYTES);
+            checkSize(32L);
             Unsafe.getUnsafe().putLong(appendAddress, value.getLong0());
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES, value.getLong1());
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES * 2, value.getLong2());
             Unsafe.getUnsafe().putLong(appendAddress + Long.BYTES * 3, value.getLong3());
-            appendAddress += Long256.BYTES;
+            appendAddress += 32L;
         }
 
         @Override
         public void putShort(short value) {
-            checkSize(Short.BYTES);
+            checkSize(2L);
             Unsafe.getUnsafe().putShort(appendAddress, value);
-            appendAddress += Short.BYTES;
+            appendAddress += 2L;
         }
 
         @Override
@@ -918,9 +920,9 @@ public class FastMap implements Map, Reopenable {
             }
 
             int len = value.length();
-            checkSize((len << 1) + Integer.BYTES);
+            checkSize(((long) len << 1) + 4L);
             Unsafe.getUnsafe().putInt(appendAddress, len);
-            appendAddress += Integer.BYTES;
+            appendAddress += 4L;
             for (int i = 0; i < len; i++) {
                 Unsafe.getUnsafe().putChar(appendAddress + ((long) i << 1), value.charAt(i));
             }
@@ -930,9 +932,9 @@ public class FastMap implements Map, Reopenable {
         @Override
         public void putStr(CharSequence value, int lo, int hi) {
             int len = hi - lo;
-            checkSize((len << 1) + Integer.BYTES);
+            checkSize(((long) len << 1) + 4L);
             Unsafe.getUnsafe().putInt(appendAddress, len);
-            appendAddress += Integer.BYTES;
+            appendAddress += 4L;
             for (int i = lo; i < hi; i++) {
                 Unsafe.getUnsafe().putChar(appendAddress + ((long) (i - lo) << 1), value.charAt(i));
             }
@@ -947,9 +949,9 @@ public class FastMap implements Map, Reopenable {
             }
 
             int len = value.length();
-            checkSize((len << 1) + Integer.BYTES);
+            checkSize(((long) len << 1) + 4L);
             Unsafe.getUnsafe().putInt(appendAddress, len);
-            appendAddress += Integer.BYTES;
+            appendAddress += 4L;
             for (int i = 0; i < len; i++) {
                 Unsafe.getUnsafe().putChar(appendAddress + ((long) i << 1), Character.toLowerCase(value.charAt(i)));
             }
@@ -959,9 +961,9 @@ public class FastMap implements Map, Reopenable {
         @Override
         public void putStrLowerCase(CharSequence value, int lo, int hi) {
             int len = hi - lo;
-            checkSize((len << 1) + Integer.BYTES);
+            checkSize(((long) len << 1) + 4L);
             Unsafe.getUnsafe().putInt(appendAddress, len);
-            appendAddress += Integer.BYTES;
+            appendAddress += 4L;
             for (int i = lo; i < hi; i++) {
                 Unsafe.getUnsafe().putChar(appendAddress + ((long) (i - lo) << 1), Character.toLowerCase(value.charAt(i)));
             }
@@ -980,13 +982,13 @@ public class FastMap implements Map, Reopenable {
         }
 
         private void putVarSizeNull() {
-            checkSize(4);
+            checkSize(4L);
             Unsafe.getUnsafe().putInt(appendAddress, TableUtils.NULL_LEN);
-            appendAddress += Integer.BYTES;
+            appendAddress += 4L;
         }
 
         @Override
-        void copyFromRawKey(long srcPtr, int srcSize) {
+        void copyFromRawKey(long srcPtr, long srcSize) {
             checkSize(srcSize);
             Vect.memcpy(appendAddress, srcPtr, srcSize);
             appendAddress += srcSize;