The core idea is to use a specialized allocator for aggregate functions with additional state. That's to amortize the cost of frequent alloc/free calls. The allocator is thread-safe, supports best-effort free operation, and frees the memory fully only when it's closed. Considering that the aggregate functions grow their state as the power of 2, the max theoretical memory overhead should be around 2x (that's only with very active data structure to thread migration). See `GroupByAllocator` for more details.

Migrated aggregate functions use new flyweight off-heap data structures backed by the allocator and store the data structure pointers in the GROUP BY map. The data structures can be found in the `GroupByCharSink`, `GroupByLongHashSet` and `GroupByIntHashSet` classes.

Includes parallel GROUP BY migration for `min(str)`/`max(str)` functions and `count_distinct()` function for LONG, INT and IPv4 types.

Also includes the following:
* Fixes broken designated timestamp push-down when parallel GROUP BY steals nested filter.

ClickBench results on my 4c/8t machine (hot runs):
```sql
-- before 33.5s
-- after 10s
-- allocated 1528MB
SELECT * FROM (SELECT REGEXP_REPLACE(Referer, '^https?://(?:www\.)?([^/]+)/.*$', '$1') AS k, AVG(length(Referer)) AS l, COUNT(*) AS c, MIN(Referer) FROM hits WHERE Referer IS NOT NULL GROUP BY k) WHERE c > 100000 ORDER BY l DESC LIMIT 25;

-- before 9.2s
-- after 2.6s
-- allocated 1021MB
SELECT RegionID, count_distinct(UserID) AS u FROM hits GROUP BY RegionID ORDER BY u DESC LIMIT 10;

-- before 10.5s
-- after 3s
-- allocated 1014MB
SELECT RegionID, SUM(AdvEngineID), COUNT(*) AS c, AVG(ResolutionWidth), count_distinct(UserID) FROM hits GROUP BY RegionID ORDER BY c DESC LIMIT 10;

-- before 500ms
-- after 250ms
-- allocated 50MB
SELECT MobilePhoneModel, count_distinct(UserID) AS u FROM hits WHERE MobilePhoneModel IS NOT NULL GROUP BY MobilePhoneModel ORDER BY u DESC LIMIT 10;

-- before 530ms
-- after 270ms
-- allocated 52MB
SELECT MobilePhone, MobilePhoneModel, count_distinct(UserID) AS u FROM hits WHERE MobilePhoneModel IS NOT NULL GROUP BY MobilePhone, MobilePhoneModel ORDER BY u DESC LIMIT 10;

-- before 7.7s
-- after 1.8s
-- allocated 2094MB
SELECT SearchPhrase, count_distinct(UserID) AS u FROM hits WHERE SearchPhrase IS NOT NULL GROUP BY SearchPhrase ORDER BY u DESC LIMIT 10;

-- before 8s
-- after 1.2s
-- allocated 3MB
SELECT SearchPhrase, MIN(URL), MIN(Title), COUNT(*) AS c, count_distinct(UserID) FROM hits WHERE Title LIKE '%Google%' AND URL NOT LIKE '%.google.%' AND SearchPhrase IS NOT NULL GROUP BY SearchPhrase ORDER BY c DESC LIMIT 10;
```

Microbenchmark results:
```
Benchmark                                           Mode  Cnt  Score    Error  Units
GroupByLongHashSetBenchmark.baseline                avgt    3  0.005 ±  0.001  us/op
GroupByLongHashSetBenchmark.testFastMap             avgt    3  0.099 ±  0.003  us/op
GroupByLongHashSetBenchmark.testGroupByLongHashSet  avgt    3  0.026 ±  0.001  us/op
```