@ideoma 2024-01-04
#### [PR Coverage check]

:heart_eyes: **pass** : 216 / 225 (96.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/cairo/map/ShardedMapCursor.java|0|3|00.00%|
|:large_blue_circle:|io/questdb/cairo/CairoConfigurationWrapper.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MaxStrGroupByFunction.java|18|19|94.74%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/MinStrGroupByFunction.java|18|19|94.74%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/GroupByCharSink.java|42|44|95.45%|
|:large_blue_circle:|io/questdb/std/IntList.java|30|31|96.77%|
|:large_blue_circle:|io/questdb/PropServerConfiguration.java|6|6|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/SampleByInterpolateRecordCursorFactory.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/groupby/GroupByAllocator.java|59|59|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/GroupByFunction.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByAtom.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByRecordCursorFactory.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/cairo/DefaultCairoConfiguration.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByUtils.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/PropertyKey.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/AsyncGroupByRecordCursorFactory.java|4|4|100.00%|
|:large_blue_circle:|io/questdb/std/MemoryTag.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/std/LongList.java|9|9|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/AbstractNoRecordSampleByCursor.java|3|3|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/groupby/GroupByNotKeyedRecordCursorFactory.java|8|8|100.00%|